LayoutTests/http/tests/cache/disk-cache/disk-cache-302-status-code.html

<script src="/js-test-resources/js-test-pre.js"></script>
<script src="resources/cache-test.js"></script>
<body>
<script>

var tests =
[
 { responseHeaders: {'Status': '302', 'Location': '/', 'Cache-control': 'max-age=0' }, includeBody: false },
 { responseHeaders: {'Status': '302', 'Location': '/', 'Cache-control': 'max-age=100' }, includeBody: false },
];

description("Test that responses with HTTP status code 302 are not cached");

debug("running " + tests.length + " tests");
debug("");

runTests(tests);

</script>
<script src="/js-test-resources/js-test-post.js"></script>
