LayoutTests/http/tests/cache/disk-cache/disk-cache-302-status-code-expected.txt

Test that responses with HTTP status code 302 are not cached

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


running 2 tests

response headers: {"Status":"302","Location":"/","Cache-control":"max-age=0"}
response source: Network

response headers: {"Status":"302","Location":"/","Cache-control":"max-age=100"}
response source: Network

PASS successfullyParsed is true

TEST COMPLETE
